+++ requirements.txt
@@ -1,0 +1,21 @@
+certifi==2020.4.5.1
+chardet==3.0.4
+Click==7.0
+decorator==4.4.1
+Flask==1.1.1
+Flask-Cors==3.0.8
+idna==2.9
+itsdangerous==1.1.0
+Jinja2==2.11.1
+MarkupSafe==1.1.1
+neo4j==1.7.6
+neobolt==1.7.16
+neotime==1.7.4
+networkx==2.4
+numpy==1.18.2
+pytz==2019.3
+requests==2.23.0
+scipy==1.4.1
+six==1.14.0
+urllib3==1.25.8
+Werkzeug==1.0.0
